# Faster imhist